resources/assets/model_test_apps/object-detection-model-file/pytorch/requirements.txt
gradio
ultralytics
Pillow
opencv-python
torch